Isolate ModelOutput from pt and tf
