+++ src/transformers/modeling_bert.py
@@ -30,13 +30,12 @@ from torch.nn import CrossEntropyLoss, MSELoss
 
 from .activations import gelu, gelu_new, swish
 from .configuration_bert import BertConfig
-from .file_utils import add_code_sample_docstrings, add_start_docstrings, add_start_docstrings_to_callable
+from .file_utils import add_code_sample_docstrings, add_start_docstrings, add_start_docstrings_to_callable, ModelOutput
 from .modeling_outputs import (
     CausalLMOutput,
     EncoderOutput,
     EncoderOutputWithPooling,
     MaskedLMOutput,
-    ModelOutput,
     MultipleChoiceModelOutput,
     NextSentencePredictorOutput,
     QuestionAnsweringModelOutput,
